Renamed sdbUtils.py to utils.py
