+++ src/toil/jobStores/aws/utils.py
